refactor(monitor): extract StallDetector from session monitor (#4249 step 1)
- New src/stall-detector.ts (443 lines): focused stall detection class
- Detects 5 stall types: jsonl, permission, unknown, extended, extended_working
- StallDetectorConfig + StallDetectorDeps interfaces for dependency injection
- Stall recovery with retryWithJitter (fire-and-forget)
- monitor.ts reduced from 1165 to 890 lines
- Backward-compat getters on SessionMonitor for external consumers
- Zero behavior changes — pure delegation refactoring
- Gate: tsc ✅, build ✅, 5652 tests ✅
